+++ server/tests/project/libs/testmodule/plugins/__init__.py
@@ -1,0 +1,6 @@
+from inmanta.plugins import plugin
+
+
+@plugin
+def plugin_symbol_test() -> "int":
+    return 0