+++ server/tests/test_smoke.py
@@ -1,11 +1,13 @@
 import json
 import logging
 import socket
+from typing import AsyncIterator, Dict, List, Optional
 
 import pytest
 import os
 
 from tornado.iostream import IOStream
+from inmantals import lsp_types
 from inmantals.jsonrpc import JsonRpcServer
 from inmantals.server import InmantaLSHandler
 from tornado.tcpclient import TCPClient
@@ -14,7 +16,7 @@ from tornado.tcpclient import TCPClient
 class JsonRPC(object):
     def __init__(self, ios: IOStream) -> None:
         self.ios = ios
-        self.rpqnr = 0
+        self.rpqnr: int = 0
 
     async def write(self, body):
         body = body.encode("utf-8")
@@ -24,8 +26,8 @@ class JsonRPC(object):
         await self.ios.write(header)
         await self.ios.write(body)
 
-    async def call(self, method, **kwargs):
-        body = {}
+    async def call(self, method, **kwargs) -> int:
+        body: Dict[str, object] = {}
 
         ident = self.rpqnr
         self.rpqnr += 1
@@ -70,11 +72,13 @@ class JsonRPC(object):
     async def assert_one(self, id):
         result = json.loads(await self.read_one())
         assert result["id"] == id
+        if "error" in result:
+            assert False, "error in response %s: %s" % (id, result["error"])
         return result["result"]
 
 
 @pytest.fixture
-async def server(event_loop):
+async def server(event_loop) -> AsyncIterator[JsonRpcServer]:
     server = JsonRpcServer(InmantaLSHandler)
     server.listen(6352)
     yield server
@@ -82,11 +86,25 @@ async def server(event_loop):
 
 
 @pytest.fixture
-async def client(server):
+async def client(server) -> AsyncIterator[JsonRPC]:
     ios = await TCPClient().connect("127.0.0.1", 6352, af=socket.AddressFamily.AF_INET)
     client = JsonRPC(ios)
     yield client
     client.ios.close()
+
+
+async def initialize_project(client: JsonRPC, project: str, client_capabilities: Optional[Dict[str, object]] = None) -> None:
+    """
+    Initializes the server with the basic_test project.
+    """
+    if client_capabilities is None:
+        client_capabilities = {}
+    path = os.path.join(os.path.dirname(__file__), "project")
+    ret = await client.call("initialize", rootPath=path, rootUri=f"file://{path}", capabilities=client_capabilities)
+    await client.assert_one(ret)
+
+    ret = await client.call("initialized")
+    await client.assert_one(ret)
 
 
 async def assert_lnr(client):
@@ -130,7 +148,7 @@ async def test_connection(client, caplog):
     caplog.set_level(logging.DEBUG)
 
     path = os.path.join(os.path.dirname(__file__), "project")
-    ret = await client.call("initialize", rootPath=path, rootUri=f"file://{path}")
+    ret = await client.call("initialize", rootPath=path, rootUri=f"file://{path}", capabilities={})
     result = await client.assert_one(ret)
     assert result == {
         "capabilities": {
@@ -143,6 +161,7 @@ async def test_connection(client, caplog):
             },
             "definitionProvider": True,
             "referencesProvider": True,
+            "workspaceSymbolProvider": {"workDoneProgress": False}
         }
     }
 
@@ -165,3 +184,96 @@ async def test_connection(client, caplog):
 
     ret = await client.call("exit")
     await client.assert_one(ret)
+
+
+def test_lsp_type_serialization() -> None:
+    """
+    LSP spec names are camel case while Python conventions are to use snake case.
+    """
+
+    class MyLspType(lsp_types.LspModel):
+        snake_case_name: int
+        optional: Optional[int]
+
+    spec_compatible: Dict = {"snakeCaseName": 0}
+
+    v1 = MyLspType(snake_case_name=0)
+    v2 = MyLspType(snakeCaseName=0)
+    v3 = MyLspType.parse_obj(spec_compatible)
+
+    for v in [v1, v2, v3]:
+        assert v.dict() == spec_compatible
+
+
+@pytest.mark.timeout(5)
+@pytest.mark.asyncio
+async def test_symbol_provider(client: JsonRPC) -> None:
+    ret: int
+    result: object
+
+    await initialize_project(client, "project")
+
+    ret = await client.call("workspace/symbol", query="symbol")
+    result = await client.assert_one(ret)
+    assert isinstance(result, list)
+    symbol_info: List[lsp_types.SymbolInformation] = [lsp_types.SymbolInformation.parse_obj(symbol) for symbol in result]
+
+    project_dir: str = os.path.abspath(os.path.join(os.path.dirname(__file__), "project"))
+    uri_main: str = "file://%s" % os.path.join(project_dir, "main.cf")
+    testmodule_dir: str = os.path.join(project_dir, "libs", "testmodule")
+    uri_testmodule_model: str = "file://%s" % os.path.join(testmodule_dir, "model", "_init.cf")
+    uri_testmodule_plugins: str = "file://%s" % os.path.join(testmodule_dir, "plugins", "__init__.py")
+
+    assert symbol_info == [
+        lsp_types.SymbolInformation(
+            name="__config__::my_symbol_test_type",
+            kind=lsp_types.SymbolKind.Class,
+            location=lsp_types.Location(
+                uri=uri_main,
+                range=lsp_types.Range(
+                    start=lsp_types.Position(line=8, character=8), end=lsp_types.Position(line=8, character=27)
+                ),
+            ),
+        ),
+        lsp_types.SymbolInformation(
+            name="testmodule::SymbolTest",
+            kind=lsp_types.SymbolKind.Class,
+            location=lsp_types.Location(
+                uri=uri_testmodule_model,
+                range=lsp_types.Range(
+                    start=lsp_types.Position(line=0, character=7), end=lsp_types.Position(line=0, character=17)
+                ),
+            ),
+        ),
+        lsp_types.SymbolInformation(
+            name="testmodule::symbolTest",
+            kind=lsp_types.SymbolKind.Constructor,
+            location=lsp_types.Location(
+                uri=uri_testmodule_model,
+                range=lsp_types.Range(
+                    start=lsp_types.Position(line=4, character=15), end=lsp_types.Position(line=4, character=25)
+                ),
+            ),
+        ),
+        lsp_types.SymbolInformation(
+            name="testmodule::plugin_symbol_test",
+            kind=lsp_types.SymbolKind.Function,
+            location=lsp_types.Location(
+                uri=uri_testmodule_plugins,
+                range=lsp_types.Range(
+                    start=lsp_types.Position(line=4, character=0), end=lsp_types.Position(line=4, character=1)
+                ),
+            ),
+        ),
+        lsp_types.SymbolInformation(
+            name="symbol",
+            kind=lsp_types.SymbolKind.Field,
+            location=lsp_types.Location(
+                uri=uri_testmodule_model,
+                range=lsp_types.Range(
+                    start=lsp_types.Position(line=1, character=0), end=lsp_types.Position(line=1, character=0)
+                ),
+            ),
+            container_name="testmodule::SymbolTest",
+        ),
+    ]