+++ server/CHANGELOG.md
@@ -1,6 +1,7 @@
 # v 1.0.0 (?)
 Changes in this release:
 - Added language server to client diagnostics notifications to highlight errors (#18)
+- Added symbol provider (#34)
 
 # v 0.2.1 (2019-10-31)